requirement.txt
python-dotenv
openai
pandas
streamlit
azure-identity